# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Assesify API
  version: "1.0.0"
  description: >
    Hand-authored OpenAPI schema describing the current Assesify Flask API
    surface (spec.md §4). Used as the source for `openapi-typescript` to
    generate frontend TS types (spec.md §7.3). Reflects the API as it exists
    today, not an aspirational redesign — kept in sync manually since the
    backend does not (yet) generate this from code.
servers:
  - url: /api/v1
    description: Versioned API root (also mounted at /api for legacy routes)
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Question:
      type: object
      required: [id, stem, qtype, options, difficulty, competency_tag]
      properties:
        id:
          type: integer
        stem:
          type: string
        qtype:
          type: string
        options:
          type: array
          items:
            type: object
        difficulty:
          type: number
        competency_tag:
          type: string
        correct_keys:
          type: array
          items:
            type: string
        explanation:
          type: string
          nullable: true
    Quiz:
      type: object
      properties:
        id:
          type: integer
        lesson_id:
          type: integer
        questions:
          type: array
          items:
            $ref: "#/components/schemas/Question"
        created_at:
          type: string
          format: date-time
...
    LegacySubmitQuizRequest:
      type: object
      properties:
        class_id:
          type: integer
          nullable: true
        answers:
          type: array
          items:
            type: object
            properties:
              question:
                type: string
              answer:
                type: string
              is_correct:
                type: boolean
    LegacySubmitQuizResponse:
      type: object
      properties:
        message:
          type: string
        attempt_id:
          type: integer
        score:
          type: number
        health:
          type: integer
        streak:
          type: integer
        diamonds_earned:
          type: integer
...
paths:
...
  /quizzes/{quiz_id}:
    get:
      operationId: getQuiz
      parameters:
        - name: quiz_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Quiz with questions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Quiz"
...
  /quizzes/{quiz_id}/submit:
    post:
      operationId: submitLegacyQuiz
      security:
        - bearerAuth: []
      parameters:
        - name: quiz_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LegacySubmitQuizRequest"
      responses:
        "201":
          description: Submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LegacySubmitQuizResponse"
  /quizzes/generate-personalized:
    post:
      operationId: generatePersonalizedQuiz
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                topic:
                  type: string
      responses:
        "200":
          description: Personalized quiz
          content:
            application/json:
              schema:
                type: object
  /quizzes/generate-weekly-test:
    post:
      operationId: generateWeeklyTest
      security:
        - bearerAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [start_date, end_date]
              properties:
                num_questions:
                  type: integer
                  default: 10
                start_date:
                  type: string
                  format: date
                end_date:
                  type: string
                  format: date
      responses:
        "201":
          description: Weekly test generated
          content:
            application/json:
              schema:
                type: object
  /quizzes/weekly-performance:
    get:
      operationId: getWeeklyPerformance
      security:
        - bearerAuth: []
      parameters:
        - name: start_date
          in: query
          required: true
          schema:
            type: string
            format: date
        - name: end_date
          in: query
          required: true
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Weekly performance summary
          content:
            application/json:
              schema:
                type: object
...